python-sdk/langwatch/examples/migration_support/todo/legacy/pyproject.toml
[tool.poetry]
name = "legacy"
version = "0.1.0"
description = ""
authors = ["Rogério Chaves <rogeriochaves@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
langsmith = "0.0.92"
langchain = "0.0.323"
langwatch = "0.1.16"
pydantic = "1.10.15"
openai = "0.28.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
